[Ads] moved directory
