+++ CMakeLists.txt
@@ -199,7 +199,7 @@ add_subdirectory( DataHolderLib )
 add_subdirectory( FileIO )
 add_subdirectory( GeoLib )
 add_subdirectory( InSituLib )
-add_subdirectory( MaterialsLib/adsorption )
+add_subdirectory( MaterialsLib/Adsorption )
 add_subdirectory( MathLib )
 add_subdirectory( MeshLib )
 add_subdirectory( MeshGeoToolsLib )
