+++ MaterialsLib/Adsorption/reaction_sinusoidal.h
